refactor: extract OAuth utilities
