+++ src/constants.ts
@@ -1,0 +1,8 @@
+export const CLIENT_ID = '05ce4abd0fc047b4bcd512f15b3445c9';
+export const CLIENT_SECRET = 'b5d7dBEe693c4C3fa50C183A3f6570D2';
+
+export const BASE_URL = 'https://anypoint.mulesoft.com';
+export const AUTHORIZATION_ENDPOINT = `${BASE_URL}/accounts/api/v2/oauth2/authorize`;
+export const TOKEN_ENDPOINT = `${BASE_URL}/accounts/api/v2/oauth2/token`;
+export const REVOKE_ENDPOINT = `${BASE_URL}/accounts/api/v2/oauth2/revoke`;
+export const LOCAL_REDIRECT_URI = 'http://localhost:8082/callback';
